RestaurantNameGenerator/requirements.txt
streamlit
openai
langchain-community
